Refactor window management: inject windowPool into managers, remove circular dependencies, and modularize logic.
