+++ src/electron/ipcSetup.js
@@ -1,0 +1,7 @@
+const setupIpcHandlers = require('./ipcHandlers');
+
+function registerIpc(windowPool, ipcMain) {
+    setupIpcHandlers(ipcMain, windowPool);
+}
+
+module.exports = { registerIpc };
