+++ src/electron/utils/displayUtils.js
@@ -1,0 +1,18 @@
+const { screen } = require('electron');
+
+function getCurrentDisplay(window) {
+    if (!window || window.isDestroyed()) return screen.getPrimaryDisplay();
+    const windowBounds = window.getBounds();
+    const windowCenter = {
+        x: windowBounds.x + windowBounds.width / 2,
+        y: windowBounds.y + windowBounds.height / 2,
+    };
+    return screen.getDisplayNearestPoint(windowCenter);
+}
+
+function getDisplayById(displayId) {
+    const displays = screen.getAllDisplays();
+    return displays.find(d => d.id === displayId) || screen.getPrimaryDisplay();
+}
+
+module.exports = { getCurrentDisplay, getDisplayById };
